feat: back-end finalizado

+++ projetoFinal/back-end/src/routes/index.ts
@@ -3,8 +3,21 @@ import CreateUserService from "../services/CreateUserService";
 import DeleteUserService from "../services/DeleteUserService";
 import ListUserService from "../services/ListUserService";
 import UpdateUserService from "../services/UpdateUserService";
+import AuthenticateUserService from "../services/AuthenticateUserService";
+import ensureAuthenticated from "../middlewares/ensureAuthenticated";
+import AppError from "../errors/AppError";
 
 const routes = Router();
+
+routes.post("/sessions", async (request, response) => {
+    const { email, password } = request.body;
+    const authenticateUserService = new AuthenticateUserService();
+
+    const { user, token } = await authenticateUserService.execute({ email, password });
+    
+    return response.json({ user, token });
+
+});
 
 routes.get("/users", async (request, response) => {
     const { email } = request.query;
@@ -15,14 +28,17 @@ routes.get("/users", async (request, response) => {
     return response.json(users);
 });
 
+
 routes.post("/users", async (request, response) => {
     const { name, email, password } = request.body;
-
+    
     const createUserService = new CreateUserService();
     const newUser = await createUserService.execute({ name, email, password });
-
+    
     return response.json(newUser);
 });
+
+routes.use(ensureAuthenticated);
 
 routes.patch("/users/:id", async (request, response) => {
     const { id } = request.params;
@@ -36,6 +52,11 @@ routes.patch("/users/:id", async (request, response) => {
 
 routes.delete("/users/:id", async (request, response) => {
     const { id } = request.params;
+    const userId = request.user.id;
+
+    if (id === userId ){
+        throw new AppError("You can't delete your own account", 400);
+    }
 
     const deleteUserService = new DeleteUserService();
     await deleteUserService.execute(id);
